Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -18,7 +18,7 @@ int main() {
             printf("%d",result);
             break;
         case '/':
-            (float)result=a/b;
+            result=a/b;
             if(b!=0){
                 printf("%f",result);}
             else{
